extract validation methods to the ValidatorTrait
